+++ example/lib/main.dart
@@ -13,7 +13,7 @@ class MyApp extends StatelessWidget {
     return MaterialApp(
       title: 'Reorderable TabBar',
       theme: ThemeData(
-        primarySwatch: Colors.blue,
+        useMaterial3: true,
       ),
       home: const ReorderableTabBarPage(),
     );
@@ -29,9 +29,12 @@ class ReorderableTabBarPage extends StatefulWidget {
 
 extension StringExt on String {
   Text get text => Text(this);
-  Widget get tab {
-    return Tab(
-      text: this,
+  Widget tab(int index) {
+    return ReorderableDragStartListener(
+      index: index,
+      child: Tab(
+        text: "Tab $this",
+      ),
     );
   }
 }
@@ -39,17 +42,11 @@ extension StringExt on String {
 class _ReorderableTabBarPageState extends State<ReorderableTabBarPage> {
   PageController pageController = PageController();
 
-  int index = 9;
-
   List<String> tabs = [
     "1",
     "2",
     "3",
     "4",
-    "5",
-    "6",
-    "7",
-    "8",
   ];
 
   bool isScrollable = false;
@@ -68,8 +65,6 @@ class _ReorderableTabBarPageState extends State<ReorderableTabBarPage> {
               padding: const EdgeInsets.all(8.0),
               child: Center(
                 child: Switch(
-                  activeColor: Colors.white,
-                  inactiveThumbColor: Colors.grey.shade400,
                   value: tabSizeIsLabel,
                   onChanged: (s) {
                     setState(() {
@@ -83,8 +78,6 @@ class _ReorderableTabBarPageState extends State<ReorderableTabBarPage> {
               padding: const EdgeInsets.all(8.0),
               child: Center(
                 child: Switch(
-                  activeColor: Colors.white,
-                  inactiveThumbColor: Colors.grey.shade400,
                   value: isScrollable,
                   onChanged: (s) {
                     setState(() {
@@ -96,10 +89,10 @@ class _ReorderableTabBarPageState extends State<ReorderableTabBarPage> {
             ),
           ],
           bottom: ReorderableTabBar(
-            tabs: tabs.map((e) => e.tab).toList(),
+            buildDefaultDragHandles: false,
+            tabs: tabs.map((e) => e.tab(tabs.indexOf(e))).toList(),
             indicatorSize: tabSizeIsLabel ? TabBarIndicatorSize.label : null,
             isScrollable: isScrollable,
-            indicatorColor: Colors.blueGrey.shade900,
             reorderingTabBackgroundColor: Colors.black45,
             indicatorWeight: 5,
             tabBorderRadius: const BorderRadius.vertical(
@@ -115,16 +108,14 @@ class _ReorderableTabBarPageState extends State<ReorderableTabBarPage> {
         floatingActionButton: FloatingActionButton(
           child: const Icon(Icons.add),
           onPressed: () {
-            tabs.add(index.toString());
-            setState(() {
-              ++index;
-            });
+            tabs.add((tabs.length + 1).toString());
+            setState(() {});
           },
         ),
         body: TabBarView(
           children: tabs.map((e) {
             return Center(
-              child: (e + ". Page").text,
+              child: ("$e. Page").text,
             );
           }).toList(),
         ),